A handful of small changes to make it compile clearly with the following parameters:
CFLAGS="-O0 -g -ggdb -fno-common -Wall -Wdeclaration-after-statement
-Wextra -Wformat=2 -Winit-self -Winline -Wpacked
-Wp,-D_FORTIFY_SOURCE=2 -Wpointer-arith -Wlarger-than-65500
-Wmissing-declarations -Wmissing-format-attribute -Wmissing-noreturn
-Wmissing-prototypes -Wnested-externs -Wold-style-definition
-Wredundant-decls -Wsign-compare -Wstrict-aliasing=2
-Wstrict-prototypes -Wundef -Wwrite-strings"

+++ src/maclist.c
@@ -96,8 +96,8 @@ mc_maclist_set_random_vendor_from_list (mac_t *mac, card_mac_list_item_t *list)
 void
 mc_maclist_set_random_vendor (mac_t *mac, mac_type_t type)
 {
-	int  num;
-	long total;
+	unsigned long num;
+	long          total;
 
 	total = LIST_LENGTH (list_others) +
 		LIST_LENGTH (list_wireless);
